Isolate rollup build with exec
